requirements.txt
PyQt5
QDarkStyle
pylzham
zstandard